Use `u64` for `file_size`

+++ parquet/src/arrow/async_reader/metadata.rs
@@ -297,7 +297,7 @@ where
     let fetch = MetadataFetchFn(fetch);
     ParquetMetaDataReader::new()
         .with_prefetch_hint(prefetch)
-        .load_and_finish(fetch, file_size)
+        .load_and_finish(fetch, file_size as u64)
         .await
 }
